Explicit VARCHAR casting to avoid Snowflake Inference
Improve the way constants are handled, and generate the SQL as follows:

Derived column definition:

derived_columns:
    RECORD_SOURCE: “!EDGE”
 
Generated SQL:

CAST('EDGE' AS VARCHAR) AS RECORD_SOURCE
 
This forces snowflake to infer the type of RECORD_SOURCE as VARCHAR(16777216), which is the default and max.
 
We can also provide a configuration option ‘cast_constants_as` which works as follows:

{{ config(cast_constants_as=’VARCHAR’) }}
derived_columns:
    RECORD_SOURCE: “!EDGE”

Generated SQL:

CAST('EDGE' AS VARCHAR) AS RECORD_SOURCE
 
This would be the same as option 1, but it adds a layer of configurability. The config could easily be:

{{ config(cast_constants_as=’VARCHAR(255)’) }}